pyproject.toml
[tool.poetry]
name = "pyanalysis-backend"
version = "0.1.0"
description = ""
authors = ["MurVladimirOS <ekliiehxmeh@gmail.com>"]
readme = "README.md"
packages = [{include = "pyanalysis_backend"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.87.0"
uvicorn = "^0.20.0"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"